refactor: 배너 영역 정보 영역UI로 변경
